Add support for Arrays/CuArrays with CUDA ext
