+++ Project.toml
@@ -3,8 +3,17 @@ uuid = "c3c07f87-98de-43f2-a76f-835b330b2cbb"
 authors = ["CliMA Contributors <clima-software@caltech.edu>"]
 version = "0.2.0"
 
+[weakdeps]
+CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
+Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
+
+[extensions]
+MultiBroadcastFusionCUDAExt = ["CUDA", "Adapt"]
+
 [compat]
 julia = "^1.10"
+Adapt = "3, 4"
+CUDA = "5"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
