+++ README.md
@@ -54,11 +54,33 @@ With this package, we can apply `@fused` to reduce the number of reads and prese
 
 ```julia
 import MultiBroadcastFusion as MBF
+x1 = rand(3,3)
+x2 = rand(3,3)
+x3 = rand(3,3)
+x4 = rand(3,3)
+y1 = rand(3,3)
+y2 = rand(3,3)
 
-MBF.@make_type FusedMultiBroadcast
-MBF.@make_fused FusedMultiBroadcast fused
-# Now, `@fused` will call `Base.copyto!(::FusedMultiBroadcast)`. Let's define it:
-function Base.copyto!(fmb::FusedMultiBroadcast)
+# 4 reads, 2 writes
+MBF.@fused begin
+  @. y1 = x1 * x2 + x3 * x4
+  @. y2 = x1 * x3 + x2 * x4
+end
+```
+
+This is achieved by fusing the loops and inlining with the given data, resulting in the compiler being able to perform Common-SubExpression Elimination (CSE) on the memory loads.
+
+## Custom implementations
+
+Users can write custom implementations, using the `@make_type` and `@make_fused` macros, and then defining `Base.copyto!` on the type you've defined
+
+```julia
+import MultiBroadcastFusion as MBF
+
+MBF.@make_type MyFusedMultiBroadcast
+MBF.@make_fused MyFusedMultiBroadcast my_fused
+# Now, `@fused` will call `Base.copyto!(::MyFusedMultiBroadcast)`. Let's define it:
+function Base.copyto!(fmb::MyFusedMultiBroadcast)
     pairs = fmb.pairs
     destinations = map(x->x.first, pairs)
     @inbounds for i in eachindex(destinations)
@@ -76,17 +98,15 @@ y1 = rand(3,3)
 y2 = rand(3,3)
 
 # 4 reads, 2 writes
-@fused begin
+@my_fused begin
   @. y1 = x1 * x2 + x3 * x4
   @. y2 = x1 * x3 + x2 * x4
 end
 ```
 
-This is achieved by fusing the loops and inlining with the given data, resulting in the compiler being able to perform Common-SubExpression Elimination (CSE) on the memory loads.
-
 ## Writing custom macros
 
-Users can write custom macros with, for example,
+Users can also write custom macros with, for example,
 
 ```julia
 import MultiBroadcastFusion as MBF
