+++ src/collection/code_lowered_single_expression.jl
