+++ src/collection/fused_pairs.jl
