+++ src/collection/fused_pairs_flexible.jl
